+++ database/migrations/2025_06_20_142106_modify_table_program_add_donate_sum_column.php
@@ -4,17 +4,18 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('program', function (Blueprint $table) {
-            $table->integer('donate_sum')->nullable();
-            $table->dateTime('donate_sum_last_updated')->nullable();
-        });
+        if (!Schema::hasColumn('program', 'donate_sum')) {
+            Schema::table('program', function (Blueprint $table) {
+                $table->integer('donate_sum')->nullable();
+                $table->dateTime('donate_sum_last_updated')->nullable();
+            });
+        }
     }
 
     /**
@@ -22,9 +23,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('program', function (Blueprint $table) {
-            $table->dropColumn('donate_sum');
-            $table->dropColumn('donate_sum_last_updated');
-        });
+        if (Schema::hasColumn('program', 'donate_sum')) {
+            Schema::table('program', function (Blueprint $table) {
+                $table->dropColumn('donate_sum');
+                $table->dropColumn('donate_sum_last_updated');
+            });
+        }
     }
 };
