+++ database/migrations/2025_06_20_142117_modify_table_program_add_is_islami_column.php
@@ -4,16 +4,17 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('program', function (Blueprint $table) {
-            $table->tinyInteger('is_islami')->default(1)->nullable();
-        });
+        if (!Schema::hasColumn('program', 'is_islami')) {
+            Schema::table('program', function (Blueprint $table) {
+                $table->tinyInteger('is_islami')->default(1)->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +22,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('program', function (Blueprint $table) {
-            $table->dropColumn('is_islami');
-        });
+        if (Schema::hasColumn('program', 'is_islami')) {
+            Schema::table('program', function (Blueprint $table) {
+                $table->dropColumn('is_islami');
+            });
+        }
     }
 };
